requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
idna==3.4
phonenumbers==8.13.4
Pillow==9.3.0
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
